LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-step-attribute.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<input id=test type=time step=60>
<script>
description('Check step attribute change causes apperance change');
var testInput = document.getElementById('test');
var widthOfStep60 = testInput.offsetWidth;

debug('Set step to 0.001');
testInput.step = 0.001
var widthOfStep001 = testInput.offsetWidth;
shouldBeTrue('widthOfStep001 > widthOfStep60');

debug('Set step to 60');
testInput.step = 60;
var widthOfStep60Again = testInput.offsetWidth;
shouldBeTrue('widthOfStep60 == widthOfStep60Again');

debug('');
testInput.parentElement.removeChild(testInput);
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
