third_party/blink/web_tests/platform/generic/fast/forms/time-multiple-fields/time-multiple-fields-step-attribute-expected.txt

Check step attribute change causes appearance change

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Set step to 0.001
PASS widthOfStep001 > widthOfStep60 is true
Set step to 60
PASS widthOfStep60 == widthOfStep60Again is true


PASS successfullyParsed is true

TEST COMPLETE
